% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{make_bar}
\alias{make_bar}
\title{Create a bar plot}
\usage{
make_bar(data, x, y, fill, title, xlab, ylab, xlabels)
}
\arguments{
\item{data}{A data frame}

\item{x}{A variable from the data param that you want to have on the x-axis}

\item{y}{A variable from the data param that you want to have on the y-axis}

\item{fill}{The variable you the bars to fill to}

\item{title}{A title for you plot}

\item{xlab}{A label for your x-axis}

\item{ylab}{A label for you y-axis}

\item{xlabels}{Labels for the variables on your x-axis}
}
\value{
A bar ggplot
}
\description{
Creates a bar plot with a given data, x, y and fill and title/labels you want
}
\examples{
data_frame <- data.frame(x = c(1,2,3,4,5), y = c(5,6,7,8,9))
make_bar(data_frame, "x", "y", "x", "My Bar Plot", "X-axis", 
"Y-axis", xlabels = c("1", "2","3", "4", "5"))
}
