% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{make_scatter}
\alias{make_scatter}
\title{Create a scatter plot}
\usage{
make_scatter(data, x, y, title, xlab, ylab, color)
}
\arguments{
\item{data}{A data frame}

\item{x}{A variable from the data param that you want to have on the x-axis}

\item{y}{A variable from the data param that you want to have on the y-axis}

\item{title}{A title for you plot}

\item{xlab}{A label for your x-axis}

\item{ylab}{A label for you y-axis}

\item{color}{The color you want your points to be}
}
\value{
A scatter ggplot
}
\description{
Creates a scatter plot with a given data, x, y and color/title/labels of your choice
}
\examples{
data_frame <- data.frame(x = c(1,2,3,4,5), y = c(5,6,7,8,9))
make_scatter(data_frame, "x", "y", "My Scatter Plot", "X-axis", "Y-axis", "Blue")
}
