% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot made}
\usage{
save_plot(name, output_file, width)
}
\arguments{
\item{name}{the name you want your plot to be}

\item{output_file}{where the file should go}

\item{width}{the width of your plot}
}
\description{
Saves plot to a path, with a width and a name
}
\examples{
directory <- "output/bike_data.csv"
save_plot("test_plot.png", directory, 10 )
}
